Tweak lists of checkboxes in taxonomy widgets
Adds small padding on the left of each option for each dash at the
beginning of the option label when a list of checkboxes (also known as
'buttons') is used to select taxonomy terms. This:

    Option A
    -Option B
    -Option C
    --Option D
    -Option E

gets visually transformed into:

    Option A
        Option B
        Option C
            Option D
        Option E

This arrangement is achieved by adding small empty blocks in from of the
text. No additional styles are applied, markup is not changed.